+++ src/endpoints/applecrashreport.rs
@@ -1,8 +1,8 @@
-use actix_web::{error, multipart, Error, HttpMessage, HttpRequest, Json, Query, State};
+use actix_web::{error, multipart, App, Error, HttpMessage, HttpRequest, Json, Query, State};
 use futures::{compat::Stream01CompatExt, StreamExt};
 
-use crate::app::{ServiceApp, ServiceState};
 use crate::endpoints::symbolicate::SymbolicationRequestQueryParams;
+use crate::services::Service;
 use crate::types::{RequestOptions, SymbolicationResponse};
 use crate::utils::multipart::{
     read_multipart_file, read_multipart_request_options, read_multipart_sources,
@@ -10,9 +10,9 @@ use crate::utils::multipart::{
 use crate::utils::sentry::ConfigureScope;
 
 async fn handle_apple_crash_report_request(
-    state: State<ServiceState>,
+    state: State<Service>,
     params: Query<SymbolicationRequestQueryParams>,
-    request: HttpRequest<ServiceState>,
+    request: HttpRequest<Service>,
 ) -> Result<Json<SymbolicationResponse>, Error> {
     sentry::start_session();
 
@@ -53,7 +53,7 @@ async fn handle_apple_crash_report_request(
     }
 }
 
-pub fn configure(app: ServiceApp) -> ServiceApp {
+pub fn configure(app: App<Service>) -> App<Service> {
     app.resource("/applecrashreport", |r| {
         let handler = compat_handler!(handle_apple_crash_report_request, s, p, r);
         r.post().with_async(handler);
@@ -67,8 +67,8 @@ mod tests {
     use actix_web::test::TestServer;
     use reqwest::{multipart, Client, StatusCode};
 
-    use crate::app::ServiceState;
     use crate::config::Config;
+    use crate::services::Service;
     use crate::test;
     use crate::types::SymbolicationResponse;
 
@@ -76,7 +76,7 @@ mod tests {
     async fn test_basic() {
         test::setup();
 
-        let service = ServiceState::create(Config::default()).unwrap();
+        let service = Service::create(Config::default()).unwrap();
         let server = TestServer::with_factory(move || crate::server::create_app(service.clone()));
 
         let file_contents = fs::read("tests/fixtures/apple_crash_report.txt").unwrap();
@@ -104,7 +104,7 @@ mod tests {
     async fn test_unknown_field() {
         test::setup();
 
-        let service = ServiceState::create(Config::default()).unwrap();
+        let service = Service::create(Config::default()).unwrap();
         let server = TestServer::with_factory(move || crate::server::create_app(service.clone()));
 
         let file_contents = fs::read("tests/fixtures/apple_crash_report.txt").unwrap();
