+++ src/endpoints/healthcheck.rs
@@ -1,13 +1,13 @@
-use actix_web::HttpRequest;
+use actix_web::{App, HttpRequest};
 
-use crate::app::{ServiceApp, ServiceState};
+use crate::services::Service;
 
-fn healthcheck(_req: HttpRequest<ServiceState>) -> &'static str {
+fn healthcheck(_req: HttpRequest<Service>) -> &'static str {
     metric!(counter("healthcheck") += 1);
     "ok"
 }
 
-pub fn configure(app: ServiceApp) -> ServiceApp {
+pub fn configure(app: App<Service>) -> App<Service> {
     app.resource("/healthcheck", |r| {
         r.get().with(healthcheck);
     })
