+++ src/services/cficaches.rs
@@ -13,11 +13,11 @@ use symbolic::{
 };
 use thiserror::Error;
 
-use crate::actors::cacher::{CacheItemRequest, CachePath, Cacher};
-use crate::actors::objects::{
+use crate::cache::{Cache, CacheKey, CacheStatus};
+use crate::services::cacher::{CacheItemRequest, CachePath, Cacher};
+use crate::services::objects::{
     FindObject, ObjectError, ObjectHandle, ObjectMetaHandle, ObjectPurpose, ObjectsActor,
 };
-use crate::cache::{Cache, CacheKey, CacheStatus};
 use crate::sources::{FileType, SourceConfig};
 use crate::types::{
     AllObjectCandidates, ObjectFeatures, ObjectId, ObjectType, ObjectUseInfo, Scope,
