+++ src/services/objects/data_cache.rs
@@ -4,7 +4,7 @@
 //! implementing the [`CacheItemRequest`] trait for a [`FetchFileDataRequest`] which can be
 //! used with a [`Cacher`] to make a filesystem based cache.
 //!
-//! [`Cacher`]: crate::actors::cacher::Cacher
+//! [`Cacher`]: crate::services::cacher::Cacher
 
 use std::cmp;
 use std::fs;
@@ -20,8 +20,8 @@ use symbolic::common::ByteView;
 use symbolic::debuginfo::{Archive, Object};
 use tempfile::{tempfile_in, NamedTempFile};
 
-use crate::actors::cacher::{CacheItemRequest, CachePath};
 use crate::cache::{CacheKey, CacheStatus};
+use crate::services::cacher::{CacheItemRequest, CachePath};
 use crate::services::download::{DownloadStatus, ObjectFileSource};
 use crate::types::{ObjectId, Scope};
 use crate::utils::futures::BoxedFuture;
