+++ src/services/objects/meta_cache.rs
@@ -14,8 +14,8 @@ use std::sync::Arc;
 use symbolic::common::ByteView;
 use symbolic::debuginfo::Object;
 
-use crate::actors::cacher::{CacheItemRequest, CachePath, Cacher};
 use crate::cache::{CacheKey, CacheStatus};
+use crate::services::cacher::{CacheItemRequest, CachePath, Cacher};
 use crate::services::download::{ObjectFileSource, ObjectFileSourceURI};
 use crate::sources::SourceId;
 use crate::types::{ObjectFeatures, ObjectId, Scope};
